pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "remote-work-auto-apply"
version = "0.1.0"
description = "Automated job discovery and matching for remote work"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.48.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.50.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
